[#159] :truck: rename: 스티커 컴포넌트 위치 변경
