+++ client/src/components/common/diary/Sticker.tsx
